Refactored class locations
