+++ src/main/java/com/academy/project/repository/CommunityAdminAndManagerRepository.java
@@ -13,9 +13,9 @@ public interface CommunityAdminAndManagerRepository extends JpaRepository<Commun
 
     Optional<CommunityAdminAndManager> findByEmailAndIsactive(String email,Boolean isactive);
     Optional<CommunityAdminAndManager> findByCognizantIdAndIsactive(String cognizantId,Boolean isactive);
-    @Query(value = "Select * from communityadminandmanager WHERE isactive = TRUE LIMIT ?1 OFFSET ?2 ", nativeQuery = true)
+    @Query(value = "Select * from communityadminandmanager WHERE isactive = TRUE ORDER BY communityadminandmanagerid LIMIT ?1 OFFSET ?2 ", nativeQuery = true)
     List<CommunityAdminAndManager> getAllActiveCommunityAdminAndManager(Integer size, Integer offset);
-    @Query(value = "SELECT * FROM communityadminandmanager WHERE isactive = TRUE ORDER BY communityadminandmanager", nativeQuery = true)
+    @Query(value = "SELECT * FROM communityadminandmanager WHERE isactive = TRUE ORDER BY communityadminandmanagerid", nativeQuery = true)
     List<CommunityAdminAndManager> getAllIsActive();
     @Query(value = "SELECT COUNT(communityadminandmanager) FROM communityadminandmanager WHERE isactive = TRUE", nativeQuery = true)
     Long counts();
