+++ src/test/java/com/academy/project/validation/EmailValidatorTest.java
@@ -1,8 +1,7 @@
 package com.academy.project.validation;
 
 import com.academy.project.exception.InvalidInputException;
-import com.academy.project.exception.RecordNotFoundException;
-import com.academy.project.model.CommunityAdminAndManager;
+import com.academy.project.helper.Validator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
